H.Res. 150 Engrossed in House
Providing for consideration of the bill (H.R. 998) to provide for the establishment of a process for the review of rules and sets of rules, and for other purposes, and providing for consideration of the joint resolution (H.J. Res. 83) disapproving the rule submitted by the Department of Labor relating to "Clarification of Employer's Continuing Obligation to Make and Maintain an Accurate Record of Each Recordable Injury and Illness".

Version:  Engrossed in House (eh)
Date:     2017-02-28
Congress: 115

Source: https://www.govinfo.gov/content/pkg/BILLS-115hres150eh/xml/BILLS-115hres150eh.xml

Sponsored-By: C001093