H.Res. 647 Engrossed in House
Providing for consideration of the bill (H.R. 477) to amend the Securities Exchange Act of 1934 to exempt from registration brokers performing services in connection with the transfer of ownership of smaller privately held companies; providing for consideration of the bill (H.R. 3971) to amend the Truth in Lending Act and the Real Estate Settlement Procedures Act of 1974 to modify the requirements for community financial institutions with respect to certain rules relating to mortgage loans, and for other purposes; and providing for consideration of the joint resolution (H.J. Res. 123) making further continuing appropriations for fiscal year 2018, and for other purposes.

Version:  Engrossed in House (eh)
Date:     2017-12-07
Congress: 115

Source: https://www.govinfo.gov/content/pkg/BILLS-115hres647eh/xml/BILLS-115hres647eh.xml

Sponsored-By: W000810