H.Res. 971 Engrossed in House
Providing for consideration of the resolution (H. Res. 970) insisting that the Department of Justice fully comply with the requests, including subpoenas, of the Permanent Select Committee on Intelligence and the subpoena issued by the Committee on the Judiciary relating to potential violations of the Foreign Intelligence Surveillance Act by personnel of the Department of Justice and related matters.

Version:  Engrossed in House (eh)
Date:     2018-06-28
Congress: 115

Source: https://www.govinfo.gov/content/pkg/BILLS-115hres971eh/xml/BILLS-115hres971eh.xml

Sponsored-By: C001093