H.Res. 1077 Reported in House
Providing for consideration of the conference report to accompany the bill (H.R. 6157) making appropriations for the Department of Defense for the fiscal year ending September 30, 2019, and for other purposes; providing for consideration of the resolution (H. Res. 1071) recognizing that allowing illegal immigrants the right to vote devalues the franchise and diminishes the voting power of United States citizens; and providing for consideration of motions to suspend the rules.

Version:  Reported in House (rh)
Date:     2018-09-25
Congress: 115

Source: https://www.govinfo.gov/content/pkg/BILLS-115hres1077rh/xml/BILLS-115hres1077rh.xml

Sponsored-By: C001053